+++ catalogue_pipeline/transformer/transformer_miro/src/main/scala/uk/ac/wellcome/platform/transformer/miro/exceptions/ShouldNotTransformException.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.transformer.exceptions
+package uk.ac.wellcome.platform.transformer.miro.exceptions
 
 class ShouldNotTransformException(message: String)
     extends RuntimeException(message)
